pyproject.toml
[project]
name = "nutrieye"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "fastapi>=0.111.0",
    "line-bot-sdk>=3.11.0",
    "openai>=1.35.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "ruff>=0.4.8",
    "mypy>=1.10.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
